requirements.txt
pyyaml>=3.1.0
docopt>=0.6.1
fabric>=1.8
